Removed circular dependency of ScanOptions in background
