+++ src/scanner/scan-parameter-generator.ts
@@ -1,8 +1,6 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-import { ScanOptions } from './exposed-apis';
 import { AxeOptions, AxeScanContext } from './launcher';
 import { RuleSifter } from './rule-sifter';
+import { ScanOptions } from './scan-options';
 
 export class ScanParameterGenerator {
     constructor(private ruleSifter: RuleSifter) {}
